requirements.txt
# Core Bot Framework
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Vector Database
chromadb==0.4.22
sentence-transformers==2.2.2

# PDF Processing - Optimiert
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Performance & Async
numpy==1.24.3
scikit-learn==1.3.2
asyncio-throttle==1.0.2

# OCR Optimization
opencv-python==4.8.1.78